pyproject.toml
[project]
name = "goiania-noise-monitor"
version = "1.0.0"
description = "Goiânia Noise Festival ticket availability monitor"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.44",
    "httpx>=0.27",
    "python-telegram-bot>=21.0",
    "pydantic>=2.0",
    "selectolax>=0.3.21",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "ruff>=0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 110
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "C4", "RUF"]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011"]
